refactor(std): merged twistededwards and bandersnatch. IsOnCurve failing for bandersnatch
