+++ std/algebra/twistededwards/curve.go
@@ -1,231 +1,61 @@
-/*
-Copyright © 2020 ConsenSys
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 package twistededwards
 
 import (
-	"errors"
-	"math/big"
-
-	"github.com/consensys/gnark-crypto/ecc"
-	edbls12377 "github.com/consensys/gnark-crypto/ecc/bls12-377/twistededwards"
-	edbls12381 "github.com/consensys/gnark-crypto/ecc/bls12-381/twistededwards"
-	edbls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315/twistededwards"
-	edbn254 "github.com/consensys/gnark-crypto/ecc/bn254/twistededwards"
-	edbw6633 "github.com/consensys/gnark-crypto/ecc/bw6-633/twistededwards"
-	edbw6761 "github.com/consensys/gnark-crypto/ecc/bw6-761/twistededwards"
 	"github.com/consensys/gnark-crypto/ecc/twistededwards"
 	"github.com/consensys/gnark/frontend"
 )
 
-// Curve methods implemented by a twisted edwards curve inside a circuit
-type Curve interface {
-	Params() *CurveParams
-	Endo() *EndoParams
-	Add(p1, p2 Point) Point
-	Double(p1 Point) Point
-	Neg(p1 Point) Point
-	AssertIsOnCurve(p1 Point)
-	ScalarMul(p1 Point, scalar frontend.Variable) Point
-	DoubleBaseScalarMul(p1, p2 Point, s1, s2 frontend.Variable) Point
-	API() frontend.API
+// curve curve is the default twisted edwards companion curve (defined on api.Curve().Fr)
+type curve struct {
+	api    frontend.API
+	id     twistededwards.ID
+	params *CurveParams
+	endo   *EndoParams
 }
 
-// Point represent a pair of X, Y coordinates inside a circuit
-type Point struct {
-	X, Y frontend.Variable
+func (c *curve) Params() *CurveParams {
+	return c.params
 }
 
-// CurveParams twisted edwards curve parameters ax^2 + y^2 = 1 + d*x^2*y^2
-// Matches gnark-crypto curve specific params
-type CurveParams struct {
-	A, D, Cofactor, Order *big.Int
-	Base                  [2]*big.Int // base point coordinates
+func (c *curve) API() frontend.API {
+	return c.api
 }
 
-// EndoParams endomorphism parameters for the curve, if they exist
-type EndoParams struct {
-	Endo   []*big.Int
-	Lambda *big.Int
+func (c *curve) Endo() *EndoParams {
+	return c.endo
 }
 
-// Coordinates of a point on a twisted Edwards curve
-type CoordTOREFACTOR struct {
-	X, Y big.Int
+func (c *curve) Add(p1, p2 Point) Point {
+	var p Point
+	p.Add(c.api, &p1, &p2, c.params)
+	return p
 }
 
-// NewEdCurve returns an Edwards curve parameters
-func NewEdCurve(api frontend.API, id twistededwards.ID) (Curve, error) {
-	snarkCurve, err := GetSnarkCurve(id)
-	if err != nil {
-		return nil, err
+func (c *curve) Double(p1 Point) Point {
+	var p Point
+	p.Double(c.api, &p1, c.params)
+	return p
+}
+func (c *curve) Neg(p1 Point) Point {
+	var p Point
+	p.Neg(c.api, &p1)
+	return p
+}
+func (c *curve) AssertIsOnCurve(p1 Point) {
+	p1.AssertIsOnCurve(c.api, c.params)
+}
+func (c *curve) ScalarMul(p1 Point, scalar frontend.Variable) Point {
+	var p Point
+	if c.endo != nil {
+		// scalar mul glv
+		p.scalarMulGLV(c.api, &p1, scalar, c.params, c.endo)
+	} else {
+		p.ScalarMul(c.api, &p1, scalar, c.params)
 	}
-	if api.Curve() != snarkCurve {
-		return nil, errors.New("invalid curve pair; snark field doesn't match twisted edwards field")
-	}
-	params, err := GetCurveParams(id)
-	if err != nil {
-		return nil, err
-	}
-	return &jubjubCurve{api: api, params: params}, nil
+	return p
 }
-
-func GetCurveParams(id twistededwards.ID) (*CurveParams, error) {
-	var params *CurveParams
-	switch id {
-	case twistededwards.BN254:
-		params = newEdBN254()
-	case twistededwards.BLS12_377:
-		params = newEdBLS12_377()
-	case twistededwards.BLS12_381:
-		params = newEdBLS12_381()
-	case twistededwards.BLS12_381_BANDERSNATCH:
-		panic("not implemented")
-	case twistededwards.BLS24_315:
-		params = newEdBLS24_315()
-	case twistededwards.BW6_761:
-		params = newEdBW6_761()
-	case twistededwards.BW6_633:
-		params = newEdBW6_633()
-	default:
-		return nil, errors.New("unknown twisted edwards curve id")
-	}
-	return params, nil
-}
-
-// GetSnarkCurve returns the matching snark curve for a twisted edwards curve
-func GetSnarkCurve(id twistededwards.ID) (ecc.ID, error) {
-	switch id {
-	case twistededwards.BN254:
-		return ecc.BN254, nil
-	case twistededwards.BLS12_377:
-		return ecc.BLS12_377, nil
-	case twistededwards.BLS12_381:
-		return ecc.BLS12_381, nil
-	case twistededwards.BLS12_381_BANDERSNATCH:
-		return ecc.BLS12_381, nil
-	case twistededwards.BLS24_315:
-		return ecc.BLS24_315, nil
-	case twistededwards.BW6_761:
-		return ecc.BW6_761, nil
-	case twistededwards.BW6_633:
-		return ecc.BW6_633, nil
-	default:
-		return ecc.UNKNOWN, errors.New("unknown twisted edwards curve id")
-	}
-}
-
-// -------------------------------------------------------------------------------------------------
-// constructors
-
-func newCurveParams() *CurveParams {
-	return &CurveParams{
-		A:        new(big.Int),
-		D:        new(big.Int),
-		Cofactor: new(big.Int),
-		Order:    new(big.Int),
-		Base:     [2]*big.Int{new(big.Int), new(big.Int)},
-	}
-}
-
-func newEdBN254() *CurveParams {
-
-	edcurve := edbn254.GetEdwardsCurve()
-	r := newCurveParams()
-	edcurve.A.ToBigIntRegular(r.A)
-	edcurve.D.ToBigIntRegular(r.D)
-	edcurve.Cofactor.ToBigIntRegular(r.Cofactor)
-	r.Order.Set(&edcurve.Order)
-	edcurve.Base.X.ToBigIntRegular(r.Base[0])
-	edcurve.Base.Y.ToBigIntRegular(r.Base[1])
-	return r
-
-}
-
-func newEdBLS12_381() *CurveParams {
-
-	edcurve := edbls12381.GetEdwardsCurve()
-
-	r := newCurveParams()
-	edcurve.A.ToBigIntRegular(r.A)
-	edcurve.D.ToBigIntRegular(r.D)
-	edcurve.Cofactor.ToBigIntRegular(r.Cofactor)
-	r.Order.Set(&edcurve.Order)
-	edcurve.Base.X.ToBigIntRegular(r.Base[0])
-	edcurve.Base.Y.ToBigIntRegular(r.Base[1])
-	return r
-
-}
-
-func newEdBLS12_377() *CurveParams {
-
-	edcurve := edbls12377.GetEdwardsCurve()
-
-	r := newCurveParams()
-	edcurve.A.ToBigIntRegular(r.A)
-	edcurve.D.ToBigIntRegular(r.D)
-	edcurve.Cofactor.ToBigIntRegular(r.Cofactor)
-	r.Order.Set(&edcurve.Order)
-	edcurve.Base.X.ToBigIntRegular(r.Base[0])
-	edcurve.Base.Y.ToBigIntRegular(r.Base[1])
-	return r
-
-}
-
-func newEdBW6_633() *CurveParams {
-
-	edcurve := edbw6633.GetEdwardsCurve()
-
-	r := newCurveParams()
-	edcurve.A.ToBigIntRegular(r.A)
-	edcurve.D.ToBigIntRegular(r.D)
-	edcurve.Cofactor.ToBigIntRegular(r.Cofactor)
-	r.Order.Set(&edcurve.Order)
-	edcurve.Base.X.ToBigIntRegular(r.Base[0])
-	edcurve.Base.Y.ToBigIntRegular(r.Base[1])
-	return r
-
-}
-
-func newEdBW6_761() *CurveParams {
-
-	edcurve := edbw6761.GetEdwardsCurve()
-
-	r := newCurveParams()
-	edcurve.A.ToBigIntRegular(r.A)
-	edcurve.D.ToBigIntRegular(r.D)
-	edcurve.Cofactor.ToBigIntRegular(r.Cofactor)
-	r.Order.Set(&edcurve.Order)
-	edcurve.Base.X.ToBigIntRegular(r.Base[0])
-	edcurve.Base.Y.ToBigIntRegular(r.Base[1])
-	return r
-
-}
-
-func newEdBLS24_315() *CurveParams {
-
-	edcurve := edbls24315.GetEdwardsCurve()
-
-	r := newCurveParams()
-	edcurve.A.ToBigIntRegular(r.A)
-	edcurve.D.ToBigIntRegular(r.D)
-	edcurve.Cofactor.ToBigIntRegular(r.Cofactor)
-	r.Order.Set(&edcurve.Order)
-	edcurve.Base.X.ToBigIntRegular(r.Base[0])
-	edcurve.Base.Y.ToBigIntRegular(r.Base[1])
-	return r
-
+func (c *curve) DoubleBaseScalarMul(p1, p2 Point, s1, s2 frontend.Variable) Point {
+	var p Point
+	p.DoubleBaseScalarMul(c.api, &p1, &p2, s1, s2, c.params)
+	return p
 }
