+++ std/algebra/twistededwards/point.go
@@ -17,18 +17,11 @@ limitations under the License.
 package twistededwards
 
 import (
-	"math/big"
-
 	"github.com/consensys/gnark/frontend"
 )
 
-// jubjubPoint point on a twisted Edwards curve in a Snark cs
-type jubjubPoint struct {
-	Point
-}
-
 // Neg computes the negative of a point in SNARK coordinates
-func (p *jubjubPoint) Neg(api frontend.API, p1 *jubjubPoint) *jubjubPoint {
+func (p *Point) Neg(api frontend.API, p1 *Point) *Point {
 	p.X = api.Neg(p1.X)
 	p.Y = p1.Y
 	return p
@@ -36,9 +29,7 @@ func (p *jubjubPoint) Neg(api frontend.API, p1 *jubjubPoint) *jubjubPoint {
 
 // AssertIsOnCurve checks if a point is on the reduced twisted Edwards curve
 // a*x² + y² = 1 + d*x²*y².
-func (p *jubjubPoint) AssertIsOnCurve(api frontend.API, curve *CurveParams) {
-
-	one := big.NewInt(1)
+func (p *Point) AssertIsOnCurve(api frontend.API, curve *CurveParams) {
 
 	xx := api.Mul(p.X, p.X)
 	yy := api.Mul(p.Y, p.Y)
@@ -47,7 +38,7 @@ func (p *jubjubPoint) AssertIsOnCurve(api frontend.API, curve *CurveParams) {
 
 	dxx := api.Mul(xx, curve.D)
 	dxxyy := api.Mul(dxx, yy)
-	rhs := api.Add(dxxyy, one)
+	rhs := api.Add(dxxyy, 1)
 
 	api.AssertIsEqual(lhs, rhs)
 
@@ -55,7 +46,7 @@ func (p *jubjubPoint) AssertIsOnCurve(api frontend.API, curve *CurveParams) {
 
 // Add Adds two points on a twisted edwards curve (eg jubjub)
 // p1, p2, c are respectively: the point to add, a known base point, and the parameters of the twisted edwards curve
-func (p *jubjubPoint) Add(api frontend.API, p1, p2 *jubjubPoint, curve *CurveParams) *jubjubPoint {
+func (p *Point) Add(api frontend.API, p1, p2 *Point, curve *CurveParams) *Point {
 
 	// u = (x1 + y1) * (x2 + y2)
 	u1 := api.Mul(p1.X, curve.A)
@@ -86,7 +77,7 @@ func (p *jubjubPoint) Add(api frontend.API, p1, p2 *jubjubPoint, curve *CurvePar
 }
 
 // Double doubles a points in SNARK coordinates
-func (p *jubjubPoint) Double(api frontend.API, p1 *jubjubPoint, curve *CurveParams) *jubjubPoint {
+func (p *Point) Double(api frontend.API, p1 *Point, curve *CurveParams) *Point {
 
 	u := api.Mul(p1.X, p1.Y)
 	v := api.Mul(p1.X, p1.X)
@@ -109,15 +100,19 @@ func (p *jubjubPoint) Double(api frontend.API, p1 *jubjubPoint, curve *CurvePara
 // curve: parameters of the Edwards curve
 // scal: scalar as a SNARK constraint
 // Standard left to right double and add
-func (p *jubjubPoint) ScalarMul(api frontend.API, p1 *jubjubPoint, scalar frontend.Variable, curve *CurveParams) *jubjubPoint {
+func (p *Point) ScalarMul(api frontend.API, p1 *Point, scalar frontend.Variable, curve *CurveParams, endo ...*EndoParams) *Point {
+	if len(endo) == 1 && endo[0] != nil {
+		// use glv
+		return p.scalarMulGLV(api, p1, scalar, curve, endo[0])
+	}
 
 	// first unpack the scalar
 	b := api.ToBinary(scalar)
 
-	res := jubjubPoint{}
-	tmp := jubjubPoint{}
-	A := jubjubPoint{}
-	B := jubjubPoint{}
+	res := Point{}
+	tmp := Point{}
+	A := Point{}
+	B := Point{}
 
 	A.Double(api, p1, curve)
 	B.Add(api, &A, p1, curve)
@@ -150,15 +145,15 @@ func (p *jubjubPoint) ScalarMul(api frontend.API, p1 *jubjubPoint, scalar fronte
 // DoubleBaseScalarMul computes s1*P1+s2*P2
 // where P1 and P2 are points on a twisted Edwards curve
 // and s1, s2 scalars.
-func (p *jubjubPoint) DoubleBaseScalarMul(api frontend.API, p1, p2 *jubjubPoint, s1, s2 frontend.Variable, curve *CurveParams) *jubjubPoint {
+func (p *Point) DoubleBaseScalarMul(api frontend.API, p1, p2 *Point, s1, s2 frontend.Variable, curve *CurveParams) *Point {
 
 	// first unpack the scalars
 	b1 := api.ToBinary(s1)
 	b2 := api.ToBinary(s2)
 
-	res := jubjubPoint{}
-	tmp := jubjubPoint{}
-	sum := jubjubPoint{}
+	res := Point{}
+	tmp := Point{}
+	sum := Point{}
 	sum.Add(api, p1, p2, curve)
 
 	n := len(b1)
