+++ std/algebra/twistededwards/scalarmul_glv.go
@@ -1,5 +1,5 @@
 /*
-Copyright © 2020 ConsenSys
+Copyright © 2022 ConsenSys Software Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package bandersnatch
+package twistededwards
 
 import (
 	"errors"
@@ -26,106 +26,17 @@ import (
 	"github.com/consensys/gnark/frontend"
 )
 
-// Point point on a twisted Edwards curve in a Snark cs
-type Point struct {
-	X, Y frontend.Variable
-}
-
-// Set a point to a point
-func (p *Point) Set(api frontend.API, p1 *Point) *Point {
-	p.X = p1.X
-	p.Y = p1.Y
-	return p
-}
-
-// Neg computes the negative of a point in SNARK coordinates
-func (p *Point) Neg(api frontend.API, p1 *Point) *Point {
-	p.X = api.Neg(p1.X)
-	p.Y = p1.Y
-	return p
-}
-
-// MustBeOnCurve checks if a point is on the reduced twisted Edwards curve
-// a*x² + y² = 1 + d*x²*y².
-func (p *Point) MustBeOnCurve(api frontend.API, curve EdCurveTOREFACTOR_BANDERSNATCH) {
-
-	one := big.NewInt(1)
-
-	xx := api.Mul(p.X, p.X)
-	yy := api.Mul(p.Y, p.Y)
-	axx := api.Mul(xx, &curve.A)
-	lhs := api.Add(axx, yy)
-
-	dxx := api.Mul(xx, &curve.D)
-	dxxyy := api.Mul(dxx, yy)
-	rhs := api.Add(dxxyy, one)
-
-	api.AssertIsEqual(lhs, rhs)
-
-}
-
-// Add Adds two points on a twisted edwards curve (eg jubjub)
-// p1, p2, c are respectively: the point to add, a known base point, and the parameters of the twisted edwards curve
-func (p *Point) Add(api frontend.API, p1, p2 *Point, curve EdCurveTOREFACTOR_BANDERSNATCH) *Point {
-
-	// u = (x1 + y1) * (x2 + y2)
-	u1 := api.Mul(p1.X, &curve.A)
-	u1 = api.Sub(p1.Y, u1)
-	u2 := api.Add(p2.X, p2.Y)
-	u := api.Mul(u1, u2)
-
-	// v0 = x1 * y2
-	v0 := api.Mul(p2.Y, p1.X)
-
-	// v1 = x2 * y1
-	v1 := api.Mul(p2.X, p1.Y)
-
-	// v2 = d * v0 * v1
-	v2 := api.Mul(&curve.D, v0, v1)
-
-	// x = (v0 + v1) / (1 + v2)
-	p.X = api.Add(v0, v1)
-	p.X = api.DivUnchecked(p.X, api.Add(1, v2))
-
-	// y = (u + a * v0 - v1) / (1 - v2)
-	p.Y = api.Mul(&curve.A, v0)
-	p.Y = api.Sub(p.Y, v1)
-	p.Y = api.Add(p.Y, u)
-	p.Y = api.DivUnchecked(p.Y, api.Sub(1, v2))
-
-	return p
-}
-
-// Double doubles a points in SNARK coordinates
-func (p *Point) Double(api frontend.API, p1 *Point, curve EdCurveTOREFACTOR_BANDERSNATCH) *Point {
-
-	u := api.Mul(p1.X, p1.Y)
-	v := api.Mul(p1.X, p1.X)
-	w := api.Mul(p1.Y, p1.Y)
-
-	n1 := api.Mul(2, u)
-	av := api.Mul(v, &curve.A)
-	n2 := api.Sub(w, av)
-	d1 := api.Add(w, av)
-	d2 := api.Sub(2, d1)
-
-	p.X = api.DivUnchecked(n1, d1)
-	p.Y = api.DivUnchecked(n2, d2)
-
-	return p
-}
-
 // phi endomorphism √-2 ∈ 𝒪₋₈
 // (x,y) → λ × (x,y) s.t. λ² = -2 mod Order
-func (p *Point) phi(api frontend.API, p1 *Point, curve EdCurveTOREFACTOR_BANDERSNATCH) *Point {
+func (p *Point) phi(api frontend.API, p1 *Point, curve *CurveParams, endo *EndoParams) *Point {
 
 	xy := api.Mul(p1.X, p1.Y)
 	yy := api.Mul(p1.Y, p1.Y)
 	f := api.Sub(1, yy)
-	f = api.Mul(f, curve.endo1)
-	g := api.Add(yy, curve.endo0)
-	g = api.Mul(g, curve.endo0)
-	h := api.Sub(yy, curve.endo0)
+	f = api.Mul(f, endo.Endo[1])
+	g := api.Add(yy, endo.Endo[0])
+	g = api.Mul(g, endo.Endo[0])
+	h := api.Sub(yy, endo.Endo[0])
 
 	p.X = api.DivUnchecked(f, xy)
 	p.Y = api.DivUnchecked(g, h)
@@ -179,7 +90,7 @@ func init() {
 // curve: parameters of the Edwards curve
 // scal: scalar as a SNARK constraint
 // Standard left to right double and add
-func (p *Point) ScalarMul(api frontend.API, p1 *Point, scalar frontend.Variable, curve EdCurveTOREFACTOR_BANDERSNATCH) *Point {
+func (p *Point) scalarMulGLV(api frontend.API, p1 *Point, scalar frontend.Variable, curve *CurveParams, endo *EndoParams) *Point {
 	// the hints allow to decompose the scalar s into s1 and s2 such that
 	// s1 + λ * s2 == s mod Order,
 	// with λ s.t. λ² = -2 mod Order.
@@ -192,7 +103,7 @@ func (p *Point) ScalarMul(api frontend.API, p1 *Point, scalar frontend.Variable,
 	s1, s2 := sd[0], sd[1]
 
 	// -s1 + λ * s2 == s + k*Order
-	api.AssertIsEqual(api.Sub(api.Mul(s2, &curve.lambda), s1), api.Add(scalar, api.Mul(&curve.Order, sd[2])))
+	api.AssertIsEqual(api.Sub(api.Mul(s2, endo.Lambda), s1), api.Add(scalar, api.Mul(curve.Order, sd[2])))
 
 	// Normally s1 and s2 are of the max size sqrt(Order) = 128
 	// But in a circuit, we force s1 to be negative by rounding always above.
@@ -204,7 +115,7 @@ func (p *Point) ScalarMul(api frontend.API, p1 *Point, scalar frontend.Variable,
 
 	var res, _p1, p2, p3, tmp Point
 	_p1.Neg(api, p1)
-	p2.phi(api, p1, curve)
+	p2.phi(api, p1, curve, endo)
 	p3.Add(api, &_p1, &p2, curve)
 
 	res.X = api.Lookup2(b1[n-1], b2[n-1], 0, _p1.X, p2.X, p3.X)
@@ -214,37 +125,6 @@ func (p *Point) ScalarMul(api frontend.API, p1 *Point, scalar frontend.Variable,
 		res.Double(api, &res, curve)
 		tmp.X = api.Lookup2(b1[i], b2[i], 0, _p1.X, p2.X, p3.X)
 		tmp.Y = api.Lookup2(b1[i], b2[i], 1, _p1.Y, p2.Y, p3.Y)
-		res.Add(api, &res, &tmp, curve)
-	}
-
-	p.X = res.X
-	p.Y = res.Y
-
-	return p
-}
-
-// DoubleBaseScalarMul computes s1*P1+s2*P2
-// where P1 and P2 are points on a twisted Edwards curve
-// and s1, s2 scalars.
-func (p *Point) DoubleBaseScalarMul(api frontend.API, p1, p2 *Point, s1, s2 frontend.Variable, curve EdCurveTOREFACTOR_BANDERSNATCH) *Point {
-
-	// first unpack the scalars
-	b1 := api.ToBinary(s1)
-	b2 := api.ToBinary(s2)
-
-	res := Point{}
-	tmp := Point{}
-	sum := Point{}
-	sum.Add(api, p1, p2, curve)
-
-	n := len(b1)
-	res.X = api.Lookup2(b1[n-1], b2[n-1], 0, p1.X, p2.X, sum.X)
-	res.Y = api.Lookup2(b1[n-1], b2[n-1], 1, p1.Y, p2.Y, sum.Y)
-
-	for i := n - 2; i >= 0; i-- {
-		res.Double(api, &res, curve)
-		tmp.X = api.Lookup2(b1[i], b2[i], 0, p1.X, p2.X, sum.X)
-		tmp.Y = api.Lookup2(b1[i], b2[i], 1, p1.Y, p2.Y, sum.Y)
 		res.Add(api, &res, &tmp, curve)
 	}
