+++ std/hints.go
@@ -2,9 +2,9 @@ package std
 
 import (
 	"github.com/consensys/gnark/backend/hint"
-	"github.com/consensys/gnark/std/algebra/bandersnatch"
 	"github.com/consensys/gnark/std/algebra/sw_bls12377"
 	"github.com/consensys/gnark/std/algebra/sw_bls24315"
+	"github.com/consensys/gnark/std/algebra/twistededwards"
 	"github.com/consensys/gnark/std/math/bits"
 )
 
@@ -13,7 +13,7 @@ func GetHints() []hint.Function {
 	return []hint.Function{
 		sw_bls24315.DecomposeScalar,
 		sw_bls12377.DecomposeScalar,
-		bandersnatch.DecomposeScalar,
+		twistededwards.DecomposeScalar,
 		bits.NTrits,
 		bits.NNAF,
 		bits.IthBit,
